@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "tsc && webpack",
+    "build": "tsc && webpack && node_modules/.bin/node-sass sass/index.scss dist/style.css",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -18,21 +18,23 @@
   },
   "homepage": "https://github.com/nrobinson295/shopping-list#readme",
   "dependencies": {
-    "@angular/common": "^6.0.6",
-    "@angular/compiler": "^6.0.6",
-    "@angular/core": "^6.0.6",
-    "@angular/forms": "^6.0.6",
-    "@angular/platform-browser": "^6.0.6",
-    "@angular/platform-browser-dynamic": "^6.0.6",
+    "@angular/common": "^6.0.7",
+    "@angular/compiler": "^6.0.7",
+    "@angular/core": "^6.0.7",
+    "@angular/forms": "^6.0.7",
+    "@angular/platform-browser": "^6.0.7",
+    "@angular/platform-browser-dynamic": "^6.0.7",
+    "@angular/router": "^6.0.7",
     "jquery": "^3.3.1",
     "rapidapi-connect": "0.0.6",
     "reflect-metadata": "^0.1.12",
     "rxjs": "^6.2.1",
-    "webpack": "^4.12.0",
+    "webpack": "^4.13.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@types/node": "^10.3.4",
+    "@types/node": "^10.5.0",
+    "node-sass": "^4.9.0",
     "webpack-cli": "^3.0.8"
   }
 }
