@@ -1,24 +1,40 @@
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
+import { RouterModule } from '@angular/router';
+import { BrowserModule } from '@angular/platform-browser';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { SetToArrayPipe } from './pipes/setToArray';
 import { RecipeService } from './services/recipeservice';
+import { AppComponent } from './components/AppComponent';
+import { RecipeForm } from './components/RecipeForm';
+import { Checkboxes } from './components/checkboxes';
 import { ShoppingListItem } from './components/ShoppingListItem';
 import { ShoppingForm } from './components/ShoppingForm';
-import { NgModule } from '@angular/core';
-import { AppComponent } from './components/AppComponent';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule } from '@angular/forms';
-import { SetToArrayPipe } from './pipes/setToArray';
-import { RecipeForm } from './components/RecipeForm';
-import { HttpClientModule } from '@angular/common/http';
+import { RecipeComponent } from './components/RecipeComponent';
+import { HomeComponent } from './components/HomeComponent';
 
+const ROUTES = [
+	{ path: '', component: HomeComponent },
+	{ path: 'recipe/:id', component: RecipeComponent }
+];
 @NgModule({
 	declarations: [
 		AppComponent,
 		ShoppingForm,
 		ShoppingListItem,
 		SetToArrayPipe,
-		RecipeForm
+		RecipeForm,
+		Checkboxes,
+		HomeComponent,
+		RecipeComponent
 	],
-	imports: [BrowserModule, FormsModule, HttpClientModule],
+	imports: [
+		BrowserModule,
+		FormsModule,
+		HttpClientModule,
+		RouterModule.forRoot(ROUTES, { useHash: true })
+	],
 	providers: [RecipeService],
 	bootstrap: [AppComponent],
 	schemas: []
